refactor: Rename ComparisonEconomicsItem to default name
It's meant to be used by environment and social comparison in a few commits
