+++ src/components/comparison/ComparisonEconomics.vue
@@ -1,41 +1,41 @@
 <template>
     <ComparisonTitle title="Macro-Economic Indicators" :studies="studies" />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Value added" 
         subtitle="Definition of total value added" 
         :get-value="getTotalAddedValue"
         :format="formatNumber"
         />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Share of agricultural GDP" 
         subtitle="Value chain GDP divided by agricultural GDP" 
         :get-value="(study) => study.ecoData?.macroData?.valueAddedShareAgriculturalGdp"
         :format="formatPercent"
         />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Share of national GDP" 
         subtitle="Value chain GDP divided by national GDP" 
         :get-value="(study) => study.ecoData?.macroData?.valueAddedShareNationalGdp"
         :format="formatPercent"
         />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Gini Index" 
         subtitle="-" 
         :get-value="(study) => study.ecoData?.macroData?.giniIndex"
         :format="formatPercent"
         />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Rate Of Integration" 
         subtitle="-" 
         :get-value="(study) => study.ecoData?.macroData?.rateOfIntegration"
         :format="formatPercent"
         />
-    <ComparisonEconomicsItem 
+    <ComparisonRow 
         :studies="studies" 
         title="Nominal Protection Coefficient" 
         subtitle="-" 
@@ -49,7 +49,7 @@
 import { formatPercent, formatNumber } from '@utils/format.js'
 import { getTotalAddedValue } from '@utils/economics.js'
 import ComparisonTitle from './ComparisonTitle.vue';
-import ComparisonEconomicsItem from './ComparisonEconomicsItem.vue';
+import ComparisonRow from './ComparisonRow.vue';
 const props = defineProps({
     studies: Array,
 })
