--- src/components/comparison/ComparisonEconomicsItem.vue
@@ -1,71 +1,0 @@
-<template>
-<tr class="rounded">
-        <td>
-            <div class="subtitle">{{ title }}</div>
-            <div class="definition">{{ subtitle }}</div>
-        </td>
-        <td v-for="(study, index) in studies" :key="`value_added__${study.id}`">
-            <div :class="getAddedValueClass(values[index])">
-                {{ values[index] ? format(values[index]) : '-' }}
-            </div>
-        </td>
-        <td></td>
-    </tr>
-</template>
-
-<script setup>
-import { computed } from 'vue';
-
-const props = defineProps({
-    studies: Array,
-    title: String,
-    subtitle: String,
-    getValue: Function,
-    format: Function
-})
-
-const values = computed(() => props.studies.map(study => props.getValue(study)))
-const getAddedValueClass = (value) => {
-    if (!value) {
-        return "gray"
-    }
-    if (value < 0) {
-        return "negative"
-    }
-    return "positive"
-}
-
-// one set if only values between 0 and 1
-// one set if only values positive
-// one set if values negative and positive
-
-// const getGdpClass = (value) => {
-//     if (!value) {
-//         return "gray"
-//     }
-//     if (value < 0.1) {
-//         return "light-blue"
-//     }
-//     return "dark-blue"
-// }
-
-// const getAddedValueClass = (value) => {
-//     if (!value) {
-//         return "gray"
-//     }
-//     if (value < 0) {
-//         return "light-red"
-//     }
-//     return "light-green"
-// }
-
-</script>
-
-<style scoped lang="scss">
-    .negative {
-        background-color: #ffac9e;
-    }
-    .positive {
-        background-color: #94d99d;
-    }
-</style>
